get-country-api/pre-requirements.txt
Flask==0.12
requests
py-healthcheck
